move config panel to own file
