+++ src/app-pages/reporting/batch-plotting/chart-settings.js
@@ -1,59 +1,28 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { connect } from 'redux-bundler-react';
 
 import Button from '../../../app-components/button';
+import ConfigurationPanel from './configuration-panel';
 import DeleteButton from '../../../app-components/delete-confirm';
 import Icon from '../../../app-components/icon';
-import MultiSelect from '../../../app-components/multi-select';
 import Select from '../../../app-components/select';
 import usePrevious from '../../../customHooks/usePrevious';
 
 import '../reporting.scss';
 
-const formatOptions = timeseries => (
-  timeseries.map(ts => ({
-    text: `${ts.instrument} - ${ts.name}`,
-    value: ts.id,
-  })).sort((a, b) => (
-    a.text < b.text
-      ? -1
-      : a.text > b.text
-        ? 1
-        : 0
-  ))
-);
-
-const configNameExists = (newConfigName = '', currentConfigurations = []) => {
-  const found = currentConfigurations.find(elem => (
-    newConfigName.trim() === elem.text.trim()
-  ));
-
-  return !!found;
-};
-
 const ChartSettings = connect(
-  'selectInstrumentTimeseriesItemsByRoute',
   'selectBatchPlotConfigurationsItems',
-  'selectProjectsIdByRoute',
   'selectBatchPlotConfigurationsActiveId',
-  'doBatchPlotConfigurationsSave',
   'doBatchPlotConfigurationsDelete',
   'doBatchPlotConfigurationsSetActiveId',
   ({
-    instrumentTimeseriesItemsByRoute: instrumentTimeseries,
     batchPlotConfigurationsItems,
-    projectsIdByRoute: project,
     batchPlotConfigurationsActiveId,
-    doBatchPlotConfigurationsSave,
     doBatchPlotConfigurationsDelete,
     doBatchPlotConfigurationsSetActiveId,
   }) => {
-    const [selectedTimeseries, setSelectedTimeseries] = useState([]);
-    const [newConfigName, setNewConfigName] = useState('');
-    const [inputError, setInputError] = useState('');
     const [isPanelOpen, setIsPanelOpen] = useState(false);
     const [isEditMode, setIsEditMode] = useState(false);
-    const timeseries = formatOptions(instrumentTimeseries);
     const previousConfigId = usePrevious(batchPlotConfigurationsActiveId);
 
     const configurations = batchPlotConfigurationsItems.map(config => ({
@@ -61,29 +30,11 @@ const ChartSettings = connect(
       value: config.id,
     }));
 
-    const handleSave = () => {
-      if (!newConfigName || !newConfigName.trim()) {
-        setInputError('Please provide a configuration name.');
-      } else if (configNameExists(newConfigName, configurations) && !isEditMode) {
-        setInputError('Configuration name already exists. Please use a different name.');
-      } else {
-        doBatchPlotConfigurationsSave({
-          ...isEditMode && { id: batchPlotConfigurationsActiveId },
-          name: newConfigName.trim(),
-          timeseries_id: selectedTimeseries,
-          project_id: project.projectId,
-        });
-        setIsPanelOpen(false);
-      }
-    };
-
     const handleEditClick = () => {
       const currentItem = batchPlotConfigurationsItems.find(elem => elem.id === batchPlotConfigurationsActiveId);
 
       if (currentItem) {
         setIsEditMode(true);
-        setNewConfigName(currentItem.name);
-        setSelectedTimeseries(currentItem.timeseries_id);
         setIsPanelOpen(true);
       }
     };
@@ -106,14 +57,6 @@ const ChartSettings = connect(
         doBatchPlotConfigurationsSetActiveId(val);
       }
     };
-
-    useEffect(() => {
-      if (!isPanelOpen) {
-        setNewConfigName('');
-        setSelectedTimeseries([]);
-        setInputError('');
-      }
-    }, [isPanelOpen, setNewConfigName, setSelectedTimeseries, setInputError]);
 
     return (
       <div className='d-flex justify-content-around'>
@@ -160,50 +103,12 @@ const ChartSettings = connect(
           />
         </div>
         {isPanelOpen && (
-          <div className='right-panel'>
-            <div className='input-container'>
-              <input
-                type='text'
-                className={`form-control${inputError ? ' is-invalid' : ''}`}
-                placeholder='Enter configuration name...'
-                value={newConfigName}
-                onChange={(e) => {
-                  if (inputError) setInputError('');
-                  setNewConfigName(e.target.value);
-                }}
-              />
-              {inputError && (
-                <div className='invalid-feedback'>
-                  {inputError}
-                </div>
-              )}
-            </div>
-            <div>
-              <MultiSelect
-                withSelectAllOption
-                menuClasses='dropdown-menu-right'
-                text={`Select Options (${(selectedTimeseries || []).length} selected)`}
-                options={timeseries}
-                onChange={val => setSelectedTimeseries(val)}
-                initialValues={selectedTimeseries}
-              />
-              <div className='panel-actions'>
-                <Button
-                  variant='secondary'
-                  size='small'
-                  text='Cancel'
-                  className='mr-2'
-                  handleClick={() => setIsPanelOpen(false)}
-                />
-                <Button
-                  variant='success'
-                  size='small'
-                  text='Save'
-                  handleClick={handleSave}
-                />
-              </div>
-            </div>
-          </div>
+          <ConfigurationPanel
+            isOpen={isPanelOpen}
+            isEditMode={isEditMode}
+            configurations={configurations}
+            closePanel={() => setIsPanelOpen(false)}
+          />
         )}
       </div>
     );
